audio: restructure audio hal code  [1/1]
PD#SWPL-47415

Problem:
audio input api unify for DVB/ES playback and support
mutli dvb audio playback

Solution:
Add mediasync policy process
Add micro speed support

Verify:
Verify it on ohm

+++ audio/aml_adecs/include/aml_dec_api.h
@@ -115,8 +115,8 @@ typedef struct aml_dec {
     dec_data_info_t raw_in_data;
     char *ad_data;
     int ad_size;
-    int in_frame_pts;
-    int out_frame_pts;
+    int64_t in_frame_pts;
+    int64_t out_frame_pts;
     int status;
     int frame_cnt;
     int fragment_left_size;
